requirements.txt
pillow>=10.0.0
pyyaml>=6.0.1
requests>=2.31.0
tqdm>=4.66.0
pandas>=2.0.0

# Optional runtime backends. The code lazy-imports these, so mock mode works
# even if they are not installed. Uncomment/install in Colab when needed.
# ultralytics>=8.2.0
# transformers>=4.45.0
# accelerate>=0.33.0
# torch>=2.3.0
